+++ src/OneBlinkRequestHandler.ts
@@ -6,6 +6,7 @@ import {
   OneBlinkResponse,
   FailResponse,
 } from './http-handlers/types'
+import OneBlinkStorageError from './OneBlinkStorageError'
 
 /**
  * Our own custom request handler to allow the response header which includes
@@ -40,9 +41,9 @@ export class OneBlinkRequestHandler<T>
     }
 
     const requestUrl = `${request.method} ${request.protocol}//${request.hostname}${request.path}?${new URLSearchParams(request.query as Record<string, string>).toString()}`
-    console.log('Starting storage upload request', requestUrl)
+    console.log('Starting storage request', requestUrl)
     const response = await this.oneBlinkHttpHandler.handleRequest(request)
-    console.log('Finished storage upload request', requestUrl)
+    console.log('Finished storage request', requestUrl)
 
     const oneblinkResponse = response.headers['x-oneblink-response']
     if (typeof oneblinkResponse === 'string') {
@@ -56,6 +57,20 @@ export class OneBlinkRequestHandler<T>
 
     return {
       response,
+    }
+  }
+
+  async sendS3Command<O>(sender: () => Promise<O>): Promise<O> {
+    try {
+      return await sender()
+    } catch (error) {
+      if (this.failResponse) {
+        throw new OneBlinkStorageError(this.failResponse.message, {
+          httpStatusCode: this.failResponse.statusCode,
+          originalError: error instanceof Error ? error : undefined,
+        })
+      }
+      throw error
     }
   }
 }